[DoctrineBridge] deprecate doctrine schema subscribers in favor of listeners
