+++ src/Symfony/Bridge/Doctrine/SchemaListener/RememberMeTokenProviderDoctrineSchemaSubscriber.php
@@ -11,39 +11,29 @@
 
 namespace Symfony\Bridge\Doctrine\SchemaListener;
 
-use Doctrine\ORM\Tools\Event\GenerateSchemaEventArgs;
+use Doctrine\Common\EventSubscriber;
+use Doctrine\ORM\Tools\ToolEvents;
 use Symfony\Bridge\Doctrine\Security\RememberMe\DoctrineTokenProvider;
-use Symfony\Component\Security\Http\RememberMe\PersistentRememberMeHandler;
-use Symfony\Component\Security\Http\RememberMe\RememberMeHandlerInterface;
+
+trigger_deprecation('symfony/doctrine-bridge', '6.3', 'The "%s" class is deprecated. Use "%s" instead.', RememberMeTokenProviderDoctrineSchemaSubscriber::class, RememberMeTokenProviderDoctrineSchemaListener::class);
 
 /**
  * Automatically adds the rememberme table needed for the {@see DoctrineTokenProvider}.
  *
  * @author Wouter de Jong <wouter@wouterj.nl>
+ *
+ * @deprecated since Symfony 6.3, use {@link RememberMeTokenProviderDoctrineSchemaListener} instead
  */
-final class RememberMeTokenProviderDoctrineSchemaSubscriber extends AbstractSchemaSubscriber
+final class RememberMeTokenProviderDoctrineSchemaSubscriber extends RememberMeTokenProviderDoctrineSchemaListener implements EventSubscriber
 {
-    private iterable $rememberMeHandlers;
-
-    /**
-     * @param iterable<mixed, RememberMeHandlerInterface> $rememberMeHandlers
-     */
-    public function __construct(iterable $rememberMeHandlers)
+    public function getSubscribedEvents(): array
     {
-        $this->rememberMeHandlers = $rememberMeHandlers;
-    }
-
-    public function postGenerateSchema(GenerateSchemaEventArgs $event): void
-    {
-        $connection = $event->getEntityManager()->getConnection();
-
-        foreach ($this->rememberMeHandlers as $rememberMeHandler) {
-            if (
-                $rememberMeHandler instanceof PersistentRememberMeHandler
-                && ($tokenProvider = $rememberMeHandler->getTokenProvider()) instanceof DoctrineTokenProvider
-            ) {
-                $tokenProvider->configureSchema($event->getSchema(), $connection, $this->getIsSameDatabaseChecker($connection));
-            }
+        if (!class_exists(ToolEvents::class)) {
+            return [];
         }
+
+        return [
+            ToolEvents::postGenerateSchema,
+        ];
     }
 }
